Made thrust float, as those are low int number and absolutely lack

+++ Source/1.5/CompProps/CompProps_EngineTrail.cs
@@ -7,7 +7,7 @@ namespace SaveOurShip2
 	public class CompProps_EngineTrail : CompProperties
 	{
 		//public GraphicData graphicData = new GraphicData();
-		public int thrust = 0;
+		public float thrust = 0;
 		public int fuelUse = 0;
 		public int width = 0;
 		public bool takeOff = false;
